Mark all enums as "unsigned"
When targetting x86_64-pc-windows-msvc, bindgen's generated bindings for
enums will come up as i32 rather than the u32 seen on Linux. This is
happening because llvm is returning that it is signed as per that
platform's expectations.

One way to fix this could be to adapt the code in assimp-rs to always
hard convert to u32, but that is messy and is not easily doable with
match patterns which provide no room for casting.  This code could be
rewritten to not use the match stanzas, but that seems a bit overkill to
work around this issue.

C itself doesn't have a good way to specify the type of the enum. C++11
however adds the 'class' keyword and ': type' e.g. ': unsigned'
underlying type specifyier. Since bindgen is invoking llvm under the
hood and it is willing to accept this C++11 syntax, just tag the enums
as such.